+++ baby-gru/src/components/button/MoorhenSimpleEditButton.js
@@ -872,43 +872,3 @@ export const MoorhenRigidBodyFitButton = (props) => {
     />
 }
 
-export const MoorhenAddSimpleButton = (props) => {
-    const selectRef = useRef()
-
-    const MoorhenAddSimplePanel = (props) => {
-        const molTypes = ['HOH', 'SO4', 'PO4', 'GOL', 'CIT', 'EDO', 'IOD', 'NA', 'CA']
-        return <Container>
-            <MenuList>
-                {molTypes.map(molType => {
-                    return <MenuItem key={molType} onClick={() => { props.onTypeSelectedCallback(molType) }}>{molType}</MenuItem>
-                })}
-            </MenuList>
-            <MoorhenMoleculeSelect {...props} allowAny={false} ref={props.selectRef} />
-        </Container>
-    }
-
-    const onTypeSelectedCallback = useCallback(async (value) => {
-        const selectedMolecule = props.molecules.find(molecule => molecule.molNo === parseInt(selectRef.current.value))
-        if (selectedMolecule) {
-            await selectedMolecule.addLigandOfType(value, props.glRef)
-            props.setSelectedButtonIndex(null)
-            const scoresUpdateEvent = new CustomEvent("scoresUpdate", { detail: { origin: props.glRef.current.origin, modifiedMolecule: selectedMolecule.molNo } })
-            document.dispatchEvent(scoresUpdateEvent)
-        }
-    }, [props.molecules, props.glRef])
-
-    return <MoorhenSimpleEditButton {...props}
-        toolTip="Add simple"
-        buttonIndex={props.buttonIndex}
-        selectedButtonIndex={props.selectedButtonIndex}
-        setSelectedButtonIndex={props.setSelectedButtonIndex}
-        needsMapData={false}
-        needsAtomData={true}
-        prompt={<MoorhenAddSimplePanel
-            {...props}
-            onTypeSelectedCallback={onTypeSelectedCallback}
-            selectRef={selectRef}
-        />}
-        icon={<img style={{ width: '100%', height: '100%' }} className="baby-gru-button-icon" src={`${props.urlPrefix}/baby-gru/pixmaps/atom-at-pointer.svg`} alt='add...' />}
-    />
-}
